fix: Move TypeBox override types into src so that they are available when SchemaBuilder is used as an external module.
