+++ packages/schema-builder/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "include": ["src/**/*", "global.d.ts"],
+  "include": ["src/**/*"],
   "extends": "../../tsconfig.json",
   "compilerOptions": {
     "rootDir": "src",
